requirements.txt
Django==2.2.2
psycopg2
python-dotenv
django-location-field
